Consolidate proto files
