+++ users-client/src/main/proto/spine/users/commands.proto
@@ -19,12 +19,12 @@
 
 syntax = "proto3";
 
-package spine.users.group;
+package spine.users;
 
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.users.group.command";
+option java_package = "io.spine.users.command";
 option java_outer_classname = "CommandsProto";
 option java_multiple_files = true;
 option java_generate_equals_and_hash = true;
@@ -32,7 +32,43 @@ option java_generate_equals_and_hash = true;
 import "spine/core/user_id.proto";
 import "spine/net/email_address.proto";
 import "spine/users/identifiers.proto";
-import "spine/users/group/group.proto";
+import "spine/users/user.proto";
+import "spine/users/group.proto";
+
+// Creates a new user account.
+message CreateUserAccount {
+
+    // The ID of the account.
+    core.UserId account = 1;
+
+    // The user for whom to create the account.
+    User user = 2 [(required) = true];
+}
+
+// Deletes a user account.
+message DeleteUserAccount {
+
+    // The ID of the account to delete.
+    core.UserId account = 1;
+}
+
+// Suspend an account of a user.
+//
+// After this command completes the user should not be able to access the application data
+// until the account is resumed.
+//
+message SuspendUserAccount {
+
+    // The ID of the account to suspend.
+    core.UserId account = 1;
+}
+
+// Activate previously suspended user account.
+message ResumeUserAccount {
+
+    // The ID of the account to resume.
+    core.UserId account = 1;
+}
 
 // Creates a group.
 message CreateGroup {
