--- users-client/src/main/proto/spine/users/user/commands.proto
@@ -1,71 +1,0 @@
-//
-// Copyright 2019, TeamDev. All rights reserved.
-//
-// Redistribution and use in source and/or binary forms, with or without
-// modification, must retain the above copyright notice and the following
-// disclaimer.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-
-syntax = "proto3";
-
-package spine.users.user;
-
-import "spine/options.proto";
-
-option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.users.user.command";
-option java_outer_classname = "CommandsProto";
-option java_multiple_files = true;
-option java_generate_equals_and_hash = true;
-
-import "spine/core/user_id.proto";
-import "spine/users/identifiers.proto";
-import "spine/users/profiles.proto";
-import "spine/users/user/user.proto";
-
-// Creates a new user account.
-message CreateUserAccount {
-
-    // The ID of the account.
-    core.UserId account = 1;
-
-    // The user for whom to create the account.
-    User user = 2 [(required) = true];
-}
-
-// Deletes a user account.
-message DeleteUserAccount {
-
-    // The ID of the account to delete.
-    core.UserId account = 1;
-}
-
-// Suspend an account of a user.
-//
-// After this command completes the user should not be able to access the application data
-// until the account is resumed.
-//
-message SuspendUserAccount {
-
-    // The ID of the account to suspend.
-    core.UserId account = 1;
-}
-
-// Activate previously suspended user account.
-message ResumeUserAccount {
-
-    // The ID of the account to resume.
-    core.UserId account = 1;
-}
