--- users-client/src/main/proto/spine/users/user/events.proto
@@ -1,69 +1,0 @@
-//
-// Copyright 2019, TeamDev. All rights reserved.
-//
-// Redistribution and use in source and/or binary forms, with or without
-// modification, must retain the above copyright notice and the following
-// disclaimer.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-
-syntax = "proto3";
-
-package spine.users.user;
-
-import "spine/options.proto";
-
-option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.users.user.event";
-option java_outer_classname = "EventsProto";
-option java_multiple_files = true;
-option java_generate_equals_and_hash = true;
-
-import "spine/core/user_id.proto";
-import "spine/net/internet_domain.proto";
-import "spine/net/email_address.proto";
-import "spine/users/identifiers.proto";
-import "spine/users/profiles.proto";
-import "spine/users/user/user.proto";
-
-// A new user account was created.
-message UserAccountCreated {
-
-    // The ID of the account.
-    core.UserId account = 1;
-
-    // The user for whom the account was created.
-    User user = 2 [(required) = true];
-}
-
-// An account of a user was deleted.
-message UserAccountDeleted {
-
-    // The ID of the deleted account.
-    core.UserId account = 1;
-}
-
-// An account of a user was suspended.
-message UserAccountSuspended {
-
-    // The ID of the suspended account.
-    core.UserId account = 1;
-}
-
-// A previously suspended account of a user was resumed.
-message UserAccountResumed {
-
-    // The ID of the resumed account.
-    core.UserId account = 1;
-}
