--- users-client/src/main/proto/spine/users/user/rejections.proto
@@ -1,73 +1,0 @@
-//
-// Copyright 2019, TeamDev. All rights reserved.
-//
-// Redistribution and use in source and/or binary forms, with or without
-// modification, must retain the above copyright notice and the following
-// disclaimer.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-//
-
-syntax = "proto3";
-
-package spine.users.user;
-
-import "spine/options.proto";
-
-option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.users.user.rejection";
-option java_multiple_files = false;
-option java_generate_equals_and_hash = true;
-
-import "spine/core/user_id.proto";
-import "spine/net/internet_domain.proto";
-import "spine/users/identifiers.proto";
-
-// A command requested creation of a new user account,
-// but the account with such ID already exists.
-message UserAccountAlreadyExists {
-
-    // The ID of the account.
-    core.UserId account = 1;
-}
-
-// A command requested deletion of a user account,
-// but the account is already marked as deleted.
-message UserAccountAlreadyDeleted {
-
-    // The ID of the account.
-    core.UserId account = 1;
-}
-
-// A command requested suspension of a user account,
-// but the account is already suspended.
-message UserAccountAlreadySuspended {
-
-    // The ID of the account.
-    core.UserId account = 1;
-}
-
-// An operation requested by a command is not available for an account which is marked
-// as deleted. Or, the account with such ID was never created.
-message UserAccountDoesNotExist {
-
-    // The ID that the command used to reference the account.
-    core.UserId account = 1;
-}
-
-// An operation is not available because an account with the passed ID was previously deleted.
-message UnavalableForPreviouslyDeletedAccount {
-
-    // The ID of the account which existed before.
-    core.UserId account = 1;
-}
