+++ users-google/src/test/java/io/spine/users/google/server/GoogleGroupLeftGroupTest.java
@@ -22,7 +22,7 @@ package io.spine.users.google.server;
 
 import io.spine.users.GroupId;
 import io.spine.users.google.event.GoogleGroupLeftParentGroup;
-import io.spine.users.group.command.RemoveGroupFromGroup;
+import io.spine.users.command.RemoveGroupFromGroup;
 import io.spine.users.server.UsersContextTest;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
