+++ users-google/src/test/java/io/spine/users/google/server/GroupDeletedTest.java
@@ -22,7 +22,7 @@ package io.spine.users.google.server;
 
 import io.spine.users.GroupId;
 import io.spine.users.google.event.GoogleGroupDeleted;
-import io.spine.users.group.command.DeleteGroup;
+import io.spine.users.command.DeleteGroup;
 import io.spine.users.server.UsersContextTest;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
