+++ users-google/src/test/java/io/spine/users/google/server/GroupEmailChangedTest.java
@@ -22,7 +22,7 @@ package io.spine.users.google.server;
 
 import io.spine.users.GroupId;
 import io.spine.users.google.event.GoogleGroupEmailChanged;
-import io.spine.users.group.command.ChangeGroupEmail;
+import io.spine.users.command.ChangeGroupEmail;
 import io.spine.users.server.UsersContextTest;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
