+++ users-google/src/test/java/io/spine/users/google/server/GroupJoinedGroupTest.java
@@ -22,7 +22,7 @@ package io.spine.users.google.server;
 
 import io.spine.users.GroupId;
 import io.spine.users.google.event.GoogleGroupJoinedParentGroup;
-import io.spine.users.group.command.AddGroupToGroup;
+import io.spine.users.command.AddGroupToGroup;
 import io.spine.users.server.UsersContextTest;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
