+++ users-server/src/main/java/io/spine/users/server/UserRoot.java
@@ -23,7 +23,7 @@ package io.spine.users.server;
 import io.spine.core.UserId;
 import io.spine.server.BoundedContext;
 import io.spine.server.aggregate.AggregateRoot;
-import io.spine.users.user.UserAccount;
+import io.spine.users.UserAccount;
 
 /**
  * The {@link UserAccount} aggregate root.
