+++ users-server/src/test/java/io/spine/users/server/group/ChangeGroupDescriptionTest.java
@@ -21,9 +21,9 @@
 package io.spine.users.server.group;
 
 import io.spine.users.GroupId;
-import io.spine.users.group.Group;
-import io.spine.users.group.command.ChangeGroupDescription;
-import io.spine.users.group.event.GroupDescriptionChanged;
+import io.spine.users.Group;
+import io.spine.users.command.ChangeGroupDescription;
+import io.spine.users.event.GroupDescriptionChanged;
 import io.spine.users.server.GroupCommandTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
