+++ users-server/src/test/java/io/spine/users/server/group/DeleteGroupTest.java
@@ -21,9 +21,9 @@
 package io.spine.users.server.group;
 
 import io.spine.users.GroupId;
-import io.spine.users.group.Group;
-import io.spine.users.group.command.DeleteGroup;
-import io.spine.users.group.event.GroupDeleted;
+import io.spine.users.Group;
+import io.spine.users.command.DeleteGroup;
+import io.spine.users.event.GroupDeleted;
 import io.spine.users.server.GroupCommandTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
