+++ users-server/src/test/java/io/spine/users/server/group/RenameGroupTest.java
@@ -21,9 +21,9 @@
 package io.spine.users.server.group;
 
 import io.spine.users.GroupId;
-import io.spine.users.group.Group;
-import io.spine.users.group.command.RenameGroup;
-import io.spine.users.group.event.GroupRenamed;
+import io.spine.users.Group;
+import io.spine.users.command.RenameGroup;
+import io.spine.users.event.GroupRenamed;
 import io.spine.users.server.GroupCommandTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
