+++ users-server/src/test/java/io/spine/users/server/user/UserAccountTest.java
@@ -23,12 +23,12 @@ package io.spine.users.server.user;
 import io.spine.core.UserId;
 import io.spine.testing.server.blackbox.BlackBoxContext;
 import io.spine.users.server.UsersContextTest;
-import io.spine.users.user.UserAccount;
-import io.spine.users.user.command.CreateUserAccount;
-import io.spine.users.user.command.DeleteUserAccount;
-import io.spine.users.user.event.UserAccountCreated;
-import io.spine.users.user.rejection.Rejections.UnavalableForPreviouslyDeletedAccount;
-import io.spine.users.user.rejection.Rejections.UserAccountAlreadyExists;
+import io.spine.users.UserAccount;
+import io.spine.users.command.CreateUserAccount;
+import io.spine.users.command.DeleteUserAccount;
+import io.spine.users.event.UserAccountCreated;
+import io.spine.users.rejection.Rejections.UnavalableForPreviouslyDeletedAccount;
+import io.spine.users.rejection.Rejections.UserAccountAlreadyExists;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
